+++ src/utils.js
@@ -18,32 +18,14 @@
     }).function;
   };
 
-  const throttle = (func, delay) => {
-    let timeout;
-
-    // make sure the function is not a constructor
-    return ({
-      function(...args) {
-        if (!timeout) {
-          timeout = setTimeout(() => {
-            timeout = null;
-
-            func.apply(this, args);
-          }, delay);
-        }
-      }
-    }).function;
-  };
-
   const moduleExports = {
-    debounce,
-    throttle
+    debounce
   };
 
   if (isModule) {
     module.exports = moduleExports;
   } else {
     window.crizmas = window.crizmas || {};
-    window.crizmas.funcUtils = moduleExports;
+    window.crizmas.componentUtils = moduleExports;
   }
 })();